Move User-route components into folders
