+++ app/routes/users/components/UserSettings/RemovePicture.tsx
