+++ app/routes/users/components/UserSettings/UserSettings.tsx
@@ -15,8 +15,8 @@ import SubmissionError from 'app/components/Form/SubmissionError';
 import { SubmitButton } from 'app/components/Form/SubmitButton';
 import { useCurrentUser } from 'app/reducers/auth';
 import { selectUserByUsername } from 'app/reducers/users';
-import DeleteUser from 'app/routes/users/components/DeleteUser';
-import RemovePicture from 'app/routes/users/components/RemovePicture';
+import DeleteUser from 'app/routes/users/components/UserSettings/DeleteUser';
+import RemovePicture from 'app/routes/users/components/UserSettings/RemovePicture';
 import { useIsCurrentUser } from 'app/routes/users/utils';
 import { useAppDispatch, useAppSelector } from 'app/store/hooks';
 import { Gender } from 'app/store/models/User';
